update(pt): polish transfer purchase ui
